backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain-google-genai
python-dotenv
pillow
pytesseract